Feat/34 team server integration

+++ src/utils/chatHandler.js
@@ -1,16 +1,17 @@
-const url = import.meta.env.VITE_CHAT_SERVER_ORIGIN;
+const url = import.meta.env.VITE_SERVER_ORIGIN;
 
 export function loadChat({ teamUid, teamName }) {
+  if (!teamUid || !teamName) throw new Error("Missing Property")
   return fetch(url + "/api/chat/load", {
     method: "POST",
     credentials: "include",
     headers: {
       "Content-Type": "application/json",
     },
-    body: {
+    body: JSON.stringify({
       teamUID: teamUid,
       teamName,
-    },
+    }),
   })
     .then((res) => res.json())
     .catch((error) => {
@@ -19,17 +20,18 @@ export function loadChat({ teamUid, teamName }) {
 }
 
 export function sendChat({ teamUid, teamName, message }) {
+  if (!teamUid || !teamName || !message) throw new Error("Missing Property")
   return fetch(url + "/api/chat/create", {
     method: "POST",
     credentials: "include",
     headers: {
       "Content-Type": "application/json",
     },
-    body: {
+    body: JSON.stringify({
       teamUID: teamUid,
       teamName,
       message,
-    },
+    }),
   })
     .then((res) => res.json())
     .catch((error) => {
@@ -38,18 +40,19 @@ export function sendChat({ teamUid, teamName, message }) {
 }
 
 export function editChat({ chatUid, teamUid, teamName, message }) {
+  if (!teamUid || !teamName || !chatUid || !message) throw new Error("Missing Property")
   return fetch(url + "/api/chat/edit", {
     method: "POST",
     credentials: "include",
     headers: {
       "Content-Type": "application/json",
     },
-    body: {
+    body: JSON.stringify({
       teamUID: teamUid,
       teamName,
       chatUID: chatUid,
       message,
-    },
+    }),
   })
     .then((res) => res.json())
     .catch((error) => {
@@ -58,17 +61,18 @@ export function editChat({ chatUid, teamUid, teamName, message }) {
 }
 
 export function deleteChat({ chatUid, teamUid, teamName }) {
+  if (!teamUid || !teamName || !chatUid) throw new Error("Missing Property")
   return fetch(url + "/api/chat/delete", {
     method: "DELETE",
     credentials: "include",
     headers: {
       "Content-Type": "application/json",
     },
-    body: {
+    body: JSON.stringify({
       teamUID: teamUid,
       teamName,
       chatUID: chatUid,
-    },
+    }),
   })
     .then((res) => res.json())
     .catch((error) => {
